requirements.txt
alembic==0.6.6
attrs==15.2.0
click==6.6
coverage==4.0a1
cryptography==1.2.3
Django==1.10.4
enum34==1.1.2
Flask==0.10.1
Flask-Bcrypt==0.6.0
Flask-Login==0.2.11
Flask-Migrate==1.2.0
Flask-Script==2.0.5
Flask-SQLAlchemy==1.0
Flask-Testing==0.4.2
Flask-WTF==0.10.2
google-common==0.0.1
gunicorn==19.0.0
idna==2.0
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.7.3
lockfile==0.12.2
Mako==1.0.0
MarkupSafe==0.23
protobuf==2.6.1
psycopg2==2.6.1
py-bcrypt==0.4
pyasn1==0.1.9
pyasn1-modules==0.0.7
pyOpenSSL==0.15.1
pyserial==3.0.1
python-editor==1.0.3
pyudev==0.16.1
service-identity==16.0.0
six==1.10.0
SQLAlchemy==0.9.4
Twisted==16.0.0
virtualenv==15.0.1
Werkzeug==0.9.6
WTForms==2.0.1
mysqlclient==1.3.7
